vocalAssistant/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.2.0
decorator==5.1.1
elevenlabs==0.2.19
executing==1.2.0
frozenlist==1.4.0
idna==3.4
iniconfig==2.0.0
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
multidict==6.0.4
openai==0.27.8
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.2.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.11
Pygments==2.15.1
pytest==7.4.0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
stack-data==0.6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.3
wcwidth==0.2.6
yarl==1.9.2
